feat(Utils): implement lotto profit rate calculate feature
로또 수익률 계산 기능 구현
수익률을 계산하여 소수점 둘째자리까지 반올림한 후 String형으로 반환하는
Utils.calculateProfitRate 함수 정의.
Lotto.java 내 있던 LottoWinner enum 클래스를 LottoWinner.java로 분리.
전체 수익계산을 위한 LottoWinner.calculateProfit 정의.
수익과 관련된 변수의 자료형은 long으로 선언.
