+++ src/main/java/lotto/Lotto.java
@@ -14,35 +14,6 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 
-enum LottoWinner {
-    FIRST(2_000_000_000, 0),
-    SECOND(30_000_000, 0),
-    THIRD(1_500_000, 0),
-    FOURTH(50_000, 0),
-    FIFTH(5_000, 0),
-    NONE(0, 0);
-
-    private final int prize;
-    private int count;
-
-    LottoWinner(int prize, int count) {
-        this.prize = prize;
-        this.count = count;
-    }
-    
-    public int getPrize() {
-        return prize;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public void incrementCount() {
-        this.count++;
-    }
-}
-
 public class Lotto {
     private final List<Integer> numbers;
 
@@ -98,7 +69,7 @@ public class Lotto {
         return new Lotto(numbers);
     }
 
-    public void checkLottoWin(List<Integer> winningNumbers, int bonusNumber) {
+    public int countMatchingNumbers(List<Integer> winningNumbers) {
         int correct = 0;
 
         for (int number : this.numbers) {
@@ -106,6 +77,12 @@ public class Lotto {
                 correct++;
             }
         }
+        return correct;
+    }
+
+    public void checkLottoWin(List<Integer> winningNumbers, int bonusNumber) {
+        int correct = countMatchingNumbers(winningNumbers);
+        
         if (correct == 6) {
             LottoWinner.FIRST.incrementCount();
         }
